+++ index.html
@@ -44,6 +44,8 @@
       >
     </p>
 
+    <button id="apply-polyfill">Apply Polyfill</button>
+
     <h2>Anchor Positioning via anchor() (stylesheet)</h2>
     <div style="position: relative">
       <div id="my-floating-positioning">Floating</div>
@@ -87,7 +89,9 @@
 
     <script type="module">
       import polyfill from '/src/index.ts';
-      polyfill();
+
+      const btn = document.getElementById('apply-polyfill');
+      btn.addEventListener('click', () => polyfill());
     </script>
   </body>
 </html>
