+++ tests/unit/transform.test.ts
@@ -1,7 +1,7 @@
 import { removeAnchorCSS } from '../../src/transform.js';
 import { getSampleCSS } from '../helpers.js';
 
-describe('remove anchor positioning CSS from source CSS', () => {
+describe('removeAnchorCSS', () => {
   it('parses and removes new anchor positioning CSS after transformation to JS', () => {
     const css = getSampleCSS('position-fallback');
     const result = removeAnchorCSS(css);
@@ -11,8 +11,4 @@ describe('remove anchor positioning CSS from source CSS', () => {
     expect(result).not.toContain('@position-fallback');
     expect(result).not.toContain('@try');
   });
-
-  // invalid selection
-  // different declaration combinations
-  //
 });
